requirements.txt
streamlit
crewai
langchain-groq
langchain-openai
PyPDF2
pandas
litellm
fpdf2
